+++ localtileserver/__init__.py
@@ -1,9 +1,9 @@
 # flake8: noqa: F401
 from localtileserver._version import __version__
+from localtileserver.application.utilities import get_cache_dir, make_vsi, purge_cache
 from localtileserver.client import RemoteTileClient
 from localtileserver.report import Report
 from localtileserver.server import TileClient, get_or_create_tile_client
-from localtileserver.utilities import get_cache_dir, make_vsi, purge_cache
 from localtileserver.widgets import (
     get_folium_tile_layer,
     get_leaflet_roi_controls,
