+++ localtileserver/__main__.py
@@ -6,8 +6,8 @@ import webbrowser
 import click
 
 from localtileserver.application import app
+from localtileserver.application.utilities import get_clean_filename
 from localtileserver.examples import get_data_path, get_pine_gulch_url
-from localtileserver.utilities import get_clean_filename
 
 
 @click.command()
