+++ localtileserver/application/palettes.py
