+++ tests/test_utilities.py
@@ -1,7 +1,11 @@
 import pytest
 
 from localtileserver import Report
-from localtileserver.palettes import get_palette_by_name, is_valid_palette_name, mpl_to_palette
+from localtileserver.application.palettes import (
+    get_palette_by_name,
+    is_valid_palette_name,
+    mpl_to_palette,
+)
 
 has_mpl = False
 try:
